[olarft] Reorganize data store into the data-store folder
